Placate g++-4.9 -Og
With -Og, not enough flow analysis is done to deduce that mons was
initialised in this loop condition, resulting in a bogus warning:

   files.cc:1852:75: warning: ‘mons’ may be used uninitialized in this
     function [-Wmaybe-uninitialized]

The warning didn't happen with -O0, which disables maybe-uninitialised
warnings; or with -O1 (or higher), which does enough analysis to figure
out that mons was initialised after all.

This change probably pleases PleasingFungus, too.