Make Blurry Vision cause reading to take longer
Try number two (ref 6df881ba), but this time implemented as a
delay, such that it actually takes longer for e.g. tele to kick
in after you start trying to read the scroll.

Currently takes one extra turn for each level of the mutation
(so 1+1 turns at level 1, 2+1 for level 2, 3+1 for level 3).
Possibly it would be better for the last level to give 4+1.

TODO: Would be good if there was a DELAY_ARMOUR_* style chance to
cancel out of the reading midway through, but I ran into
technical issues with the implementation - its messaging caused
the interrupt to fire immediately every time...?