Display success chance for targeted MR-checking spells
Also enabled for Banishment right now.

Making this resist more transparent has been the goal of two recent changes
to monster MR: moving from HD multipliers to fixed amounts, and displaying
that amount in monster lookup through x-v.

Hexes in particular are mostly binary success/fail outcome with monster MR
and spellpower as factors. One of these, monster MR, is fully visible
in-game as of recently, and the other is available in a slightly fuzzed
way. Both use bars instead of plain numbers for display purposes.

When implementing the chance display, I considered applying a fuzz to the
displayed chances as well, as to not "leak information". After playtesting
the fuzzed version, I revised my opinion and now believe it is much better
to present the actual not-fuzzed chances right where they belong.
For reference, "fuzzed" here means using the spellpower number corres-
ponding to the highest power bar displayed for calculating chances.

Because this certainly will come up in discussions, here are my thoughts:
Raw spellpower numbers are obfuscated through bars not because knowing
about them would break the game in several ways, but because the numbers
themselves would carry little meaning in the disconnected screen that
currently holds the power bars.

This is different in two ways from showing a chance that aims to correctly
provide players with information relevant to their choices: The number that
is calculated here has a clear and concise meaning (how likely to succeed)
and is presented right where it matters (in the targeter itself).

Of course players can, and some will, use the numbers to calculate their
spellpower - but it's not like that is impossible as of now, and more
importantly they cannot do much more with that information than what the
functionality added with this commit already provides.

Brought to compilation by |amethyst and to correctness by rwbarton, thanks!