Simplify stabbing when not using a short blade.
The extra damage that a one-handed polearm or long blade got were never
really meaningful in practice. The same is true of clubstabbing; it was a
good joke, but clubs are such a weapon of last resort (or first, in
practice) that you'd never run into it.

Everything in the removed middle tier got bumped down to the normal stab
multiplier, except for felid claws, which moved to short blade (but not
dagger) effectiveness.