Remove the book of Wizardry.
It's long since lost any semblance of thematic connectivity and has
instead become a dumping ground for spells we can't find any other good
books for, not to mention making some extremely powerful spells possibly
too common.

The only spell there not in any other book, Spellforged Servitor, is now
in the book of Power (which is no longer artificially common as of the
last commit). Every other spell is in exactly one other book; if you're
hungry for one of those spells to be common again, you can always
worship Sif Muna.