Smooth dodging penalties from armour
Wearing armour gives two separate penalties to EV. There's the AEVP penalty,
which scales with strength, armour skill, and the square of armour base EVP.
This is the component that can give you "negative" EV when in very heavy
armour. There's also a linear EVP penalty (with a formula that apparently
dates back to 4.1...?), which scales with strength, armour base EVP, and
size.

The latter penalty is the problem one - since it's capped by Dodging skill,
it makes Dodging do nothing until you "overcome" the penalty, at which
point Dodging gives EV at the same rate as if you had no armour at all.
This is misleading for new players (since it looks for all the world as
if Dodging is completely useless when in heavy armour) and rather clunky
beside.

This changes the linear penalty to divide EV from dodging rather than
subtracting from it. This makes heavy armour better at low Dodging and
worse at high Dodging; or, in other words, makes small amounts of
Dodging investment useful for heavy armour users, but high levels of
Dodging less of a priority. Which seems reasonable.

Some example values are here: http://pastebin.com/nKt9dqJH .