Warn before cancelling blurry ?blink (Grunt)
Since it wastes one (or more!) turns, when you have blurry vision.

Other cancellable scrolls (e.g. ?ew, ?ea, amnesia) could in
principle have this interface, but the turn cost isn't really
relevant for those, and I suspect the UI annoyance of having a
prompt to cancel would be greater than the benefits.

It would be nice if the blink prompt warned you *why* you might
not want to cancel.