New unrand - Gyre and Gimble (argonaut).
This version differs from the original patch in significant ways; in
fact, gameplay-wise, it's a whole new weapon. Instead of being,
effectively, a two-hander for short blades (i.e., the best in terms of
base damage that the class can provide), it's literally two strikes per
attack. I added vorpal to it because that definitely lowered damage
versus AC, and, of course, for the Jabberwocky reference.

Tile by roctavian.