Add Xom mood meter to webtiles.
This was added a few weeks ago to console and local tiles.

This isn't exactly the same, on console and local tiles the meter
turns red if Xom is getting BORED and lightred when Xom is BORED.
Unfortunately without making either an ugly hack or exposing
gift_timeout in the json this behavior can't be replicated in
webtiles; though the old colouring behavior was also not present
in webtiles so I'm guessing this isn't too big of a loss.