Don't let allies cast Animate Dead while worshipping Beogh.
There is a strange anti-synergy that begins to develop when orc wizards grow
up into orc sorcerors. They begin to animate corpses (almost) indiscriminately
and you end up with a massive army of zombies. This is problematic when Beogh
is already the god of *orcish* armies. Zombies frequently get in the way of the
pathing and LOF of your "congregation," forcing you to do tedious juggling just
to do something trivial, like change floors with your current orc pack.

Finally, if anyone wanted to do anything with Beogh resurrection this would be
a prerequisite.