Slay the undead minotaur
In labyrinth_lava_lair. Incredibly boring and tedious to fight
(non-regenerating, trapped behind lava, utterly harmless...), and
you're encouraged to do so for the loot, of course.

Replace it with a pair of lava snakes, which, in a lab, could
plausibly kill someone. (And owing to the nature of the terrain,
blocking off terrain with roughly stationary enemies is a more
interesting mechanic...)