Add a small chance of confusing actors pulled with Gell's.
It was a bit underwhelming previously; doing this should aid its utility
as a translocation intended to keep monsters away from you, as well as
add some utility as a hex.

This doesn't impact the target of the hex; they're already in enough
trouble by virtue of large collision damage if you're using this well.