Simplify debug_monspells a bit.
Removing a sign warning in the process, as well as a bounds error
that would occur if a monster book ever had NUM_MSTYPES as a value
(a bug for sure, but no reason to stomp memory).

This relies on (monster_type) 0 not being a valid monster, but so
did the old code.

Constify a few things too (but not systematically).