Clarify Shadow Creatures' description, bring back "Abyssal matter".
The previous wording implied that the monster's items were permanent
(well, at least that's what I thought when I first saw it). It also
sounded a bit like it summoned monsters that were copies of ones you
could see; hopefully this is a bit clearer.

"Abyssal matter" was changed because it apparently implied that it
should be unholy (back when the spell was unholy), but I personally
like the flavour and don't think anybody would think that if one didn't
know that the spell used to be unholy.