Support for attack flavours in forms.
Not just unarmed brands, but AF_* that can stack on top of weapons.

Lich form gets AF_DRAIN_XP; ice form gets the currently irrelevant
AF_COLD.

The attack flavours do not apply in unarmed combat so as not to conflict
with get_uc_brand(), but I can easily see the latter going away if this
change sticks.