Don't block Sublimation at 50 max MP (#9335)
We were temporarily setting MP to 50 to avoid the enough-MP checks
in spell_is_uncastable.  Instead add back just the spell's cost;
this could still be incorrect if Xom casts Sublimation for you when
you are at maxmp-2, or if an ability casts a spell but costs less
mana than the spell would.

Also remove a second version of the Sublimation check now that the one
in spell_is_uncastable works in this situation.

PF points out that we probably shouldn't be checking enough_mp in
spell_is_uncastable at all.