Add ability_slot option.
In addition to the original patch:

Allow the user to enable/disable overwriting in *_slot options.  A + in
the letter list turns on overwriting (swapping with existing letter
assignments) for the subsequent letters, while a - turns it off.
Default is on for item_slot, off for ability_slot and spell_slot.

[Committer's note: Updated and rebased by hyperactiveChipmunk, many
 thanks. Closes #162 pull request. -nfm]