Make the plain spider a genus enemy only.
They are almost exactly the same enemy as water moccasins and yellow
wasps (less the slowing effect in the latter case), and having them
appear in large numbers anywhere merely presents a popcorn threat that
we're not really looking for.

I've tried to come up with reasonable replacements for the handful of
vaults that place them, though there are still a few questionable ones
(do we still need to have the spider basket trap?).