Change the randbook formula.
It now is based on the average of all spells in the book; the old
formula seemed a bit strange, and I completely broke it during a
refactoring anyway. The rarity of the rarest spell in the book is
multiplied by 3, to avoid Shock being in a book with Tornado making it too
cheap. To be honest, I'm still not really happy with this new reordering;
the rarity of the most common book a spell is in isn't a great indicator
of its value at all. Anyone else can feel free to repaint this particular
bikeshed.