Reform Elf Tiles
The old elf tiles were bland and tepid. They didn't reflect the place
of elves in the natural order of the dungeon. These tiles reform the
essence of what it means to be elfish. Lovingly handcrafted in small
batches, these little buddies protect the awesome Christmas presents
hidden in Elf:$.