Make sure Santa gets mail forwarded from his old address.
Look for save chunks and bones under "Coc", not "Nor"; otherwise, if
a player saved in Coc and reloads for the holidays, they wind up in
a brand-new level, possibly embedded in rock or suspended over deep
water.

This change causes the same problem for people who saved in Nor, but
1. the end of the holidays will do that anyway; and 2. there are no
such saves on the servers at the moment.