Don't trap people in a minifort (theTower, #9328)
The default-mini_float behaviour doesn't actually connect up the
exit point that was found: it just verifies that at least one such
point was connected.  Still, that would have been a veto, but
currently deep water counts for the has_exits check.

This commit doesn't change either of those things, but works around
them by adding an explicit exit to the vault with @.