Fix an abyss test.
abyss_short_wait is supposed to not have monsters, and fails if any do
appear, with "A foo is nearby! xN Infinite lua loop detected, aborting."

For one, explicitly dismiss monsters on entering the abyss (rather than
just disabling spawns).  Also, just in case (for example, if we generate
a new abyss area that contains a vault), wait with . rather than 5 so we
will advance the turn counter regardless.