Remove Dith's fire suppression
Not as necessary now that all wands ID on pickup, and didn't seem
to be working anyway? (If someone can figure out how to get it
working & adds an appropriate message when it's triggered, I
wouldn't mind it returning to make use of e.g. Disc of Storms less
suicidal...)

This reverts commit 7e837e58.