Turn mon_holy_type into a bitfield.
This adds two new flags MH_UNHOLY and MH_EVIL that a mons can pick up
if it isn't something automatically unholy (demonic) or evil (undead)
but has a god or spells that taint it.

The player also picks up these flags so that is_unholy/evil/holy
are consistent with the bitfield entry at all times.