Don't allow traps in a few arrival vaults (#8760)
A teleport trap could allow the boring beetle or jelly to escape.
Not that a D:1 jelly is beyond the pale, but it seems unintentional
regardless.

You can still teleport into the boring beetle vault, but it does
place both a dungeon exit and an escape hatch down, and the beetle
is slow enough to run away from.