Move the greaterplayer colour to a pleasing shade of cyan
The greaterplayer yellow was a bit too similar similar to devteam
orange, especially when both names were bold. The number of goodplayers
who are not also greatplayers is actually small, so the visual clash was
more frequent than for greatplayer vs. greaterplayer. This shade of cyan
is reasonably distinct from the blue for greatplayer, so that all role
colours are easy to recognize.