Redo essence/purity/health sacrifice piety
Increase the piety for -Wiz and -stats based on your base stats. If the
relevant stat is higher among your stats, you get some extra piety.
-MP sacrifices get extra piety equal to half of the skill level of your
highest mp-boosting skill.

All mutations which remove a positive mutation get bonus piety equal
to the piety you get for the new negative mutation.

Overall this is an increase to piety from these lesser sacrifices, but
it is a reduction for characters for whom these were previously trivial.