refactor _get_randart_properties
This refactoring has a few goals: 1) improve the maintainability of the
code, making future refinements easier, 2) remove the odd outcome patterns
that came from checking for properties in a fixed order, 3) more or less
preserve the current level of artefact quality, 4) de-randomize the level
of certain bad properties (did you know that a randart with *Rage could
inspire rage on between 1 and 9 out of 100 attacks?)

This refactoring intentionally changes or ignores a fair amount of the
previous logic; gone are the specific frequencies and orders of
properties, for example. I've also loosened restrictions on where stealth
and magic resistance can spawn.

This also fixes a bug where see invis could not spawn on rings of
invisibility, but could spawn on rings of see invis (to no purpose).

It's possible that this could do with further refactoring. Other devs,
please let me know.