Merge add_autopickup_func and add_no_autopickup_func; add clear_autopickup_funcs.
The way add_autopickup_func now works is the following:
* it adds a function to the list chk_force_autopickup
* when considering an item for autopickup, evaluate that item with each
  function in chk_force_autopickup, in order
* if one such function returns true/false, stop and use that to determine
  whether to autopickup the item
* if none of the functions returns a boolean, continue on to other ways of
  configuring autopickup (autopickup_exceptions and autopickup item classes)

This will unfortunately break some rcfiles. In particular, if an rcfile
used add_autopickup_func to add a function that returns false for some items,
then this now means that those items will not be autopickuped (even if they
normally would be). In most cases players can change those "return false"
lines to "return" to return to the old behavior.

We currently don't seem to have any documentation for advanced Lua usage such
as this other than the example in advanced_optioneering.txt, which has been
updated.

In addition, a clear_autopickup_funcs() function was added to clear
chk_force_autopickup. This allows players to remove default autopickup Lua
(such as book autopickup on Trogites).