Refactor config loading, error handling, and runtime initialization
Previously logging was initialized in conf.py and configuration checks
requiring only the parsed data were happening in server.py. Errors were
also going through the logging interface regardless of where in the
configuration process we were.

This commit separates out the config-data-only checks so that they're
all in conf.py and moves logging initialization to server.py. Now Conf()
raises a ConfigError with a messages formatted based on the file that
was the source of the error (there can be many since we load game config
data from a directory).

The main configuration data in e.g. config.toml is now only read into
config.data via config.read() without checking or loading further data,
so the minimal information is available for any logging initialization,
chroot(), or drop of privileges. After this config.load() is called to
do the full checking and loading of additional config files.