Simplify item.subtype() clua.
Most of the stuff in this function was ancient and not used anywhere, and
the function was useless for most object types. Now it just calls
sub_type_string() (first checking whether the player knows the subtype),
except for the case of OBJ_ARMOUR. A very large number of player scripts
used the old OBJ_ARMOUR behavior (returning the armour slot name) and
would be broken by changing it, and the behavior is useful anyway.