Add a new fake language
From https://code.google.com/p/buttbot/ :

"Buttbot is a social experiment in humour and interaction, it
playfully mocks those around it. Almost deconstructionist in
nature by manipulating the context of the text it parodies,
it forces people to reexamine what has been said in a new light.

It builds on the tradition of mechanising surrealist techniques,
such as example Dadadodo by Jwz implementing the cut-up technique
made famous by William S. Burroughs (and used by Radiohead).
However instead of merely rearranging the content, buttbot aims
to augment it with the simplistic low brow humour.

The use of puerile humour is mainly to bring the conversation
back down to earth, to remind us that in our intellectual pursuits
that childish references to anatomy can bring a smile to our
faces..."

To use it, set "language = butt" in your rcfile.