Add Makefile targets to build only the WebTiles client files
This splits the dependencies of the webserver target into the targets
webserver-client and webserver-game, with the webserver itself depending
on both. Now `make webserver-client' can be used to prepare only the
files necessary for upgrading the WebTiles client, so that the server
can be upgraded without changing any game-specific WebTiles files.