Rescale the spellcasting penalty for AEVP
The penalty multiplier has been reduced from 19 to 25 and the free -0.8
AEVP bonus is removed. This means that there's no arbitrary breakpoint
past which all spellcasting penalty is removed for any armour that actually
has a penalty. Lighter armours overall are harder to cast in. Very heavy
armours are slightly easier to cast in, particularly at high strength and
Armour skill.

The motivation is to 1) remove the invisible bonus, 2) make the heavier light
armours less of a no-brainer.