Update the javascript contrib libraries
This updates all but image.js from requirejs-plugins, since that is
already the most recent version.

I've removed the specific version numbers in the local filenames, since
these version numbers can be seen in the files themselves, but use the
correct corresponding version in cdn URL for when cdn is enabled.

Since the .browser attribute is deprecated for jquery 1.9+, we use
jquery.browser.js, a compat lib. We may not actually need to be so
restrictive for disabling inflate, since recent Opera versions use blink
as rendering library and recent Safari shoudl work. I've enabled inflate
for Safari for now, since only old versions might still have a
problem. In any case further research can varify which versions of
Safari and Opera are actually support our use of inflate for binary
Websocket frame.