Adjust stat/slaying and multi-pip resistance logic
Now when a property is chosen of these types, it increments a certain amount
(1 for pips, 2 + coinflip for stats/slaying), uses up a good property,
and then has a chance to repeat the process up to a certain limit.

This also increases the number of possible good properties from 6 to 7
in order to compensate for the weakening of these properties per good
property chosen.