Adjust some uses of negative energy resistance
Mostly an attempt to get rid of uses of one_chance_in(rN), and make negative
energy effects have a smaller number of different ways of working.

Vampiric weapons and AF_VAMPIRIC are now both blocked by one level of rN,
instead of with one chance in rN. Monster vampiric draining is also blocked by
one level of rN (as player vampiric draining already was). AF_DRAIN_STAT is
no longer affected by rN at all, since it's not necromantically flavoured
and is used by things such as ancient zymes and ghost moths. The draining
effect on monsters is no longer blocked with one chance in rN, but is reduced
in duration in a vaguely similar way to how draining is reduced on players.