Remove Forceful Dismissal
In a lot of cases it'd just be more useful to spend MP on summoning more things
rather than explode your summons, but it was also thematically just not a
Summoning spell, giving the school a way to deal a bunch of direct damage.
Spells like Fulminant Prism do that sort of thing with a Summonings component
in a better way.