Remove Beoghite last-hitting (Lasty)
It should no longer matter who gets the last hit in on a monster,
unless you really want blessings to go to a specific orc.

Chances go from 2/3 for orcpals getting a blessing from a kill
& 1/10 from the player getting a kill to a uniform 1/2 chance,
very roughly matching the total number of blessings that a player
who got 2 kills for every 1 follower kill would get in the old
system. (2:1 seemed to be a fairly common ratio.)