Fix Shoals autoexplore on characters with passive mapping (3672, 7847).
Previously autoexplore in Shoals on characters with passive mapping would
usually never complete. This is more noticeable in new Shoals, which is
actually autoexplorable normally by characters without permaflight or
swimming.

What was happening is that passive mapping was calling magic_mapping(),
which was seeing that some squares had changed terrain (because of tides)
and then marking those squares as unseen. Then autoexplore would go
re-explore those squares.

Now magic_mapping() simply updates the player's map with the new terrain
for such squares if the player has seen them before (and they are left
still marked as seen). Map knowledge and feature code is complicated enough
that I'm not 100% confident that this won't cause issues with something
else, but it makes sense that mapping shouldn't be able to mark squares as
unseen.