Clear away marked cards code.
CFLAG_SEEN is able to perform just as well as CFLAG_MARKED for more
or less all purposes. CFLAG_MARKED told you what a card was when it
was on top even if it hadn't been before, but since Mark Four's removal
you could only mark cards that were on top of the deck (with ?id and
Stack Five). They could get moved lower by Nemelex wrath shuffling,
but that also cleared markings!