Give the player an ugly to use against hostile uglies.
Most of the deck of summonings no longer has any chance for hostility,
and Repulsiveness stands out in the current situation. Foxfire is the only
other one, and it almost always gives something to help the player at the
same time; so Repulsiveness can borrow a bit of that idea, and also
showcase the interaction between hostile ugly things.