Monstrous Menagerie cap: 4->3
It's generally considered extremely strong, even for a level 6
spell, and there seemed to be dev consensus between dropping the
summon cap slightly. (2 was felt to be too low, since "as a
ranged/support spell... it needs some numbers to work...")

(And, also, it's not *that* strong. Level 6 spells are allowed to
be good!)

I honestly thought this had been pushed at the time, over a month
ago, but apparently no one did?