Make butchering single-turn again.
With a delay this time, so the timing works out.

The first corpse you butcher each turn gets butchered right away instead
of getting pushed onto the delay queue like the others. I'm not sure if
this is too annoying, but it would be nontrivial to change so I'll just
leave it like this and allow playtesting.