Rework Gozag's gold distraction to use a player duration
Rather than tie it to specific piles of gold, use a duration and prop on the
player to keep track of when to distract monsters. This fixes a number of
issues with the previous implementation: players are no longer encouraged to
avoid picking up fresh gold or to kite monsters to it, and the effect can still
be triggered if the monster is exploded into chunks.

Functionally this strengthens the effect somewhat, since individual piles of
gold no longer expire one by one - the effect stacks for as long as the player
keeps refreshing the duration by killing new things.

Rather than display a status light constantly, the gold line on the HUD is
highlighted while distraction is active.