Disable Gozag Duplication
It's a very ineffective way to help out Gozag worshippers early on, since it
varies hugely in effectiveness with early lucky finds. Without a big find
(such as a good AC/EV ring, etc), it can function as an early game boost if
used on something boring like a small stack of blinking/hasting/heal wounds,
but the fact that it works on stacks at all suggests to the player that they
should wait and use it on a big stack of something exciting, even though the
additional help is unlikely to be nearly as useful by that point in the game.

On top of this, it depends very strongly on Gozag's wrath being effective and
threatening, since otherwise it's very easy for any character at all to
duplicate their early luck and then just abandon Gozag. Gozag's previous wrath
didn't achieve this at all, and the new wrath is very likely to need testing
and adjustment. Having a more thoroughly tested and effective wrath would be
a minimum pre-requisite to bringing back immediately-accessible Duplication in
some form.