Turn off autopickup of redundant rings after Sac Hand for non-octopodes
If you aren't an octopode, Sac Hand makes duplicates of all static rings
useless, so this script disables pickup on those rings if you're already
carrying one and aren't an octopode.

The macabre finger necklace is an exception, but it's uncommon enough that
I'm fine with leaving that case to manual control.