Remove rPois- as an artefact property.
The stacking for rPois with this was buggy and didn't work as intended (it
didn't handle simultaneously wearing artefacts with rPois+ and with rPois-
correctly). Even if working properly, the intended behavior was quite
unnatural to players -- displaying rPois as rPois+/- made it seem like
rPois should stack like other resists (and then be capped between rPois+
and rPois-).

Making rPois stack like this is certainly an option (and possibly a good
idea in the long run), but rPois- artefacts also haven't played out to be
particularly interesting -- rPois- is mainly relevant in early game and
some Lair branches, and players don't usually have many artefacts by then.
So this commit disables rPois- artefacts (removing rPois- from artefacts in
transferred games) and leaves rPois stacking as it was previously for
temporary sources of rPois-: having multiple sources of rPois+ is no
better than having a single source.