Bring jiyva_stat_action() more up to date with modern crawl.
Specifically, heavy armour penalties to spellcasting and dodging worked
differently when much of this function was written, and the effect of Str/Dex
on melee damage was doubled at some point. This meant that this function
was often undervaluing Str relative to Dex; hopefully it is a bit better at
this now.

Also, remove a magic EVP cutoff point that made more sense 4 years ago.