Don't word wrap in the line reader.
This behaviour change, introduced in 0.8.0-a0-6162-g597a4ca, wasn't
reflected in the redrawing and cursor positioning code. Go back to the
space-agnostic column wrapping, but not to the old wrapcprintf code,
which didn't handle double-width or even multibyte characters.

There are still bugs related to cursor positioning and redrawing when a
double-width character wraps when there was still a single space on the
line.

This doesn't solve #5554 (reproducible crash in the line reader),
but probably does make it easier to fix.