Repeat viable selections more correctly (#9498)
If you selected background first, then chose viable race, you would as
expected get a viable race for that background.  But then when repeating
that selection for the next game (with Tab), you would have a 50% chance
of first selecting a viable race, then selecting the background,
resulting in a completely random race selection.  Now, if one selection
was 'viable', do the other one first.  If both were viable, continue to
use a random order.

If the player's selection was in fact viable race first, then background,
it was already being remembered as "random race", so that case isn't
affected.