Remove a non-functional subst.
The Lua subst() call, unlike the SUBST: keyword in the des format, does
not accept a comma-separated list, but only a single substitution.  But it
appears the line wouldn't have done anything anyway, since HIJKL aren't
generated until later substs a few lines down.

Not changing to the "intended" behaviour (splitting the subst call and
moving it later) because that would be a difficulty nerf:

<st_> I didn't realise it was wrong so it can't have been too wrong :p