Wz -> Re
Rename, adjust starting stats, change description (which was
wrong before!),

'Wizard' sounds to a new player like a pure spellcasting class,
which it's not really the best suited to - something like cj or
fe is probably better for that playstyle. Reaver is a proven name
for the role.

Not totally sure about the save compat here for old reaver scores.