Remove direct damage from stat loss while already at zero
Being permaslowed, on top of the additional stat-specific penalties, is likely
to be more than enough to kill any character where this would be relevant - in
practice the damage mostly just made the mechanic a bit more complicated.

This also means it's no longer necessary to track the cause for stat loss, so
the code is simplified somewhat too.