Remove unique annotations for god gifts and pacification (dpeg, #9475)
If you convert Saint Roka his annotation follows you around even though
he's an ally. Additionally, prior to this patch, a annotation was only
removed for a pacified unique if the player saw the unique leave. This
has been changed to remove the annotation on pacification.

This has a possible edge case that I don't think can actually happen. If
the unique is pacified and then becomes hostile again before leaving out
of the player's sight the annotation won't be re-added, which could lead
to a player being surprised. If this can happen (maybe with clouds?)
it's far less frequent than an annotation remaining because the player
did not see the monster leave.

[Committer's note: removed a set of braces. -nfm]