The macabre finger can dig in your pockets even when you can't.
This fixes 9497, probably other weird things that happened with the
macabre finger and forms that meld rings but not ammulets. Specifically:

- No longer meld a ring on the amulet when melding only rings
- Permit switching rings on the end of the amulet when in such a transform.

This is a tiny buff to the amulet's interaction with some specific forms.
I'm tempted to ammend this with some updated flavour text for the amulet
suggesting just how creepy the described behavior is.