Removing the macabre finger with a ring on it is no longer super fast
Previous code would call remove_ring which should not be called recursively,
as it adjusts you.time_taken every time it is called. Now we fail (loudly)
if a player tries to remove the necklace while it has a ring on it. This
is less elegant interface-wise, but doing the right thing otherwise would
be a larger refactor.

(cherry picked from commit b642a18a849cc69654e466430a8f05870284bbc6)