Let shafts work on flying monsters, remove save based on body weight
Since shafts now work on flying players, allow shafts (as well as the shafting
effects from the stone of tremors and the Shaft card) to work on flying
monsters too. Change the weight-based save to a fixed 1/4 chance to escape for
regular dungeon shafts (the same as the current chance for a human to escape).

(cherry picked from commit 8b1fb446bf6ea29fdd2dc52bfed12850422fc898)