Simplify shop/shopping list interaction with duplicate stock (#8934)
Previously had a display bug, and didn't make sense. This still has the
wart that if a player selects to purchase one of a duplicate group the
whole group is removed from the shopping list, but it at least makes
sense over the old behavior (which was to keep the item selected, but
display nothing, and buggily accumulate nonsense total cost display).

Commiter's note: I simplified a conditional block, moved an ASSERT to a
better location, and reformatted.

(cherry picked from commit 555aecbca65a055f58a209afdc28d17532d81093)