Mulch all the ammo. Mulch it now!
The plan is to remove ammo mulching (aside from throwing nets), and to
multiply the amount of ammo spawned by the old mulch rate. It might be
a little high, but I am leaving enemy ammo spawns exactly the same, so
that monsters get the same number of shots. I expect this to lower the
amount of ammo available overall, and I welcome that change as a way to
differentiate ranged combat a little further.