Don't repeatedly time out monster torpor (tedric)
It was timing out as usual on the slowed monster's turn, then
reactivating on the snail's turn.  Use similar logic to what already
existed in timeout_enchantments (which only applies when the player
leaves and returns to the level): if the monster is under the torpor
effect, don't slow it at all.

There still could be message spam if something breaks LOS between
the snail and the torpees.

(cherry picked from commit 9e78d70e11d573fd5bcc77bd4072399af1e48d9a)