Let ctrl-f see books' last spells' schools (#9534)
This does let books match the words "Level" and "Type", but that has
been the case for a while, anyway, since an extra newline was added.  If
we want to remove that header, we should do so in a way that is less
fragile than "skip the first two formatted_string ops".

(cherry picked from commit de53a1839f29e53f04516953afff46aff945b982)