Be more careful with duplicate ghost annotations (rast)
If you had two ghosts with the same name, combo, and XL on two different
levels, the annotation tracker would treat them as one monster: seeing
one (or killing it) would remove the annotation for the other.

Now, only remove annotations for ghosts on the current level---since
they can never change levels, we don't have to worry about the same
ghost showing up twice.

There's still a problem if duplicate ghosts appear on the same level:
only one will be annotated, and killing it will remove the annotation
until the other is seen.